+++ example/models/bar.ts
@@ -1,5 +1,9 @@
+import React from 'react';
+
 export default function() {
+  const [count, setCount] = React.useState(0);
   return {
-    description: 'bar',
+    description: count,
+    setCount,
   };
 }