+++ packages/plugin-access/src/utils/getAccessProviderContent.ts
@@ -1,51 +1,21 @@
-export default function() {
+import { utils } from 'umi';
+import { join } from 'path';
+
+export default function(util: typeof utils) {
   return `\
 import React, { useMemo } from 'react';
-import { useModel } from '../core/umiExports';
 import { IRoute } from 'umi';
-import accessFactory from '@/access';
+import { useModel } from '../core/umiExports';
+import accessFactory from '../../access';
 import AccessContext, { AccessInstance } from './context';
-
-const _routes = require('../core/routes').routes;
+import { traverseModifyRoutes } from '${util.winPath(
+    join(__dirname, '..', 'utils', 'runtimeUtil'),
+  )}';
 
 type Routes = IRoute[];
 
-function traverseModifyRoutes(routes: Routes, access: AccessInstance = {} as AccessInstance) {
-  const resultRoutes: Routes = [].concat(routes as any);
-  const notHandledRoutes: Routes = [];
-
-  notHandledRoutes.push(...resultRoutes);
-
-  for (let i = 0; i < notHandledRoutes.length; i++) {
-    const currentRoute = notHandledRoutes[i];
-    let currentRouteAccessible = typeof currentRoute.unaccessible === 'boolean' ? !currentRoute.unaccessible : true;
-    if (currentRoute && currentRoute.access) {
-      if (typeof currentRoute.access !== 'string') {
-        throw new Error('[plugin-access]: "access" field set in "' + currentRoute.path + '" route should be a string.');
-      }
-      const accessProp = access[currentRoute.access];
-      if (typeof accessProp === 'function') {
-        currentRouteAccessible = accessProp(currentRoute)
-      } else if (typeof accessProp === 'boolean') {
-        currentRouteAccessible = accessProp;
-      }
-      currentRoute.unaccessible = !currentRouteAccessible;
-    }
-
-    if (currentRoute.routes || currentRoute.childRoutes) {
-      const childRoutes: Routes = currentRoute.routes || currentRoute.childRoutes;
-      if (!Array.isArray(childRoutes)) {
-        continue;
-      }
-      childRoutes.forEach(childRoute => { childRoute.unaccessible = !currentRouteAccessible }); // Default inherit from parent route
-      notHandledRoutes.push(...childRoutes);
-    }
-  }
-
-  return resultRoutes;
-}
-
 interface Props {
+  routes: Routes;
   children: React.ReactNode;
 }
 
@@ -57,13 +27,13 @@ const AccessProvider: React.FC<Props> = props => {
   const { children } = props;
   const { initialState } = useModel('@@initialState');
 
-  const access = useMemo(() => accessFactory(initialState as any), [initialState]);
+  const access: AccessInstance = useMemo(() => accessFactory(initialState as any), [initialState]);
 
   if (process.env.NODE_ENV === 'development' && (access === undefined || access === null)) {
     console.warn('[plugin-access]: the access instance created by access.ts(js) is nullish, maybe you need check it.');
   }
 
-  _routes.splice(0, _routes.length, ...traverseModifyRoutes(_routes, access));
+  props.routes.splice(0, props.routes.length, ...traverseModifyRoutes(props.routes, access));
 
   return React.createElement(
     AccessContext.Provider,