+++ packages/plugin-dva/src/fixtures/normal/.umirc.ts
@@ -1,13 +1,9 @@
 
 export default {
-  history: {
-    type: 'memory',
-    options: {
-      initialEntries: ['/'],
-    },
-  },
-  mountElementId: '',
   routes: [
     { path: '/', component: 'index' },
+  ],
+  plugins: [
+    require.resolve('../../'),
   ],
 }