+++ packages/plugin-dva/src/fixtures/with-immer/.umirc.ts
@@ -1,16 +1,12 @@
 
 export default {
-  history: {
-    type: 'memory',
-    options: {
-      initialEntries: ['/'],
-    },
-  },
-  mountElementId: '',
   routes: [
     { path: '/', component: 'index' },
   ],
   dva: {
     immer: true,
   },
+  plugins: [
+    require.resolve('../../'),
+  ],
 }