+++ packages/plugin-dva/src/index.test.ts
@@ -1,6 +1,7 @@
 import { join } from 'path';
-import { Service, utils } from 'umi';
-import { render, fireEvent, getByText, cleanup } from '@testing-library/react';
+import { utils } from 'umi';
+import { fireEvent, getByText, cleanup } from '@testing-library/react';
+import { generateTmp, render } from '@umijs/test-utils';
 
 const fixtures = join(__dirname, 'fixtures');
 
@@ -8,20 +9,8 @@ afterEach(cleanup);
 
 test('normal', async () => {
   const cwd = join(fixtures, 'normal');
-  const service = new Service({
-    cwd,
-    plugins: [require.resolve('./')],
-  });
-  await service.run({
-    name: 'g',
-    args: {
-      _: ['g', 'tmp'],
-    },
-  });
-
-  const { container } = render(
-    require(join(cwd, '.umi-test', 'umi.ts')).default,
-  );
+  await generateTmp({ cwd });
+  const { container } = render({ cwd });
   expect(container.innerHTML).toEqual(
     '<div><h1 class="title">Page index foo 0</h1></div>',
   );
@@ -29,20 +18,8 @@ test('normal', async () => {
 
 test('page models', async () => {
   const cwd = join(fixtures, 'page-models');
-  const service = new Service({
-    cwd,
-    plugins: [require.resolve('./')],
-  });
-  await service.run({
-    name: 'g',
-    args: {
-      _: ['g', 'tmp'],
-    },
-  });
-
-  const { container } = render(
-    require(join(cwd, '.umi-test', 'umi.ts')).default,
-  );
+  await generateTmp({ cwd });
+  const { container } = render({ cwd });
   expect(container.innerHTML).toEqual(
     '<div><h1 class="title">Page index foo 0 bar 1</h1></div>',
   );
@@ -50,20 +27,8 @@ test('page models', async () => {
 
 test('with-immer', async () => {
   const cwd = join(fixtures, 'with-immer');
-  const service = new Service({
-    cwd,
-    plugins: [require.resolve('./')],
-  });
-  await service.run({
-    name: 'g',
-    args: {
-      _: ['g', 'tmp'],
-    },
-  });
-
-  const { container } = render(
-    require(join(cwd, '.umi-test', 'umi.ts')).default,
-  );
+  await generateTmp({ cwd });
+  const { container } = render({ cwd });
   expect(container.innerHTML).toEqual(
     '<div><h1 class="title">Page index foo 0</h1><button>add</button></div>',
   );