+++ packages/plugin-initial-state/src/utils/getExportContent.tsx
@@ -1,6 +1,7 @@
-export default (
-  modelPath: string,
-) => `import { InitialState as InitialStateType } from '../${modelPath}';
+// @ts-ignore
+export default (modelPath: string) => `
+// @ts-ignore
+import { InitialState as InitialStateType } from '../${modelPath}';
 
 export type InitialState = InitialStateType;
 export const __PLUGIN_INITIAL_STATE = 1;