+++ packages/plugin-locale/fixtures/multiple-ext/config/config.js
@@ -4,6 +4,7 @@ export default {
   ],
   locale: {
     baseNavigator: false,
+    useLocalStorage: false,
     default: 'en-US',
   },
   singular: true,