+++ packages/plugin-locale/src/index.ts
@@ -1,16 +1,20 @@
 import { IApi } from 'umi';
 import { join, dirname } from 'path';
-import { readFileSync, existsSync } from 'fs';
+import { readFileSync } from 'fs';
 import {
   IGetLocaleFileListResult,
+  IAddAntdLocales,
   getLocaleList,
   isNeedPolyfill,
   exactLocalePaths,
+  getMomentLocale,
+  getAntdLocale,
 } from './utils';
 
 interface ILocaleConfig {
   default?: string;
   baseNavigator?: boolean;
+  useLocalStorage?: boolean;
   /** title 开启国际化 */
   title?: boolean;
   antd?: boolean;
@@ -20,7 +24,7 @@ interface ILocaleConfig {
 export default (api: IApi) => {
   const {
     paths,
-    utils: { Mustache, winPath },
+    utils: { Mustache, winPath, lodash },
   } = api;
 
   api.describe({
@@ -28,11 +32,13 @@ export default (api: IApi) => {
     config: {
       default: {
         baseNavigator: true,
+        useLocalStorage: true,
         baseSeparator: '-',
       },
       schema(joi) {
         return joi.object({
           default: joi.string(),
+          useLocalStorage: joi.boolean(),
           baseNavigator: joi.boolean(),
           title: joi.boolean(),
           antd: joi.boolean(),
@@ -50,12 +56,26 @@ export default (api: IApi) => {
     }));
   }
 
-  const getList = (): IGetLocaleFileListResult[] => {
+  const addAntdLocales: IAddAntdLocales = async args =>
+    await api.applyPlugins({
+      key: 'addAntdLocales',
+      type: api.ApplyPluginsType.add,
+      initialValue: [
+        `antd/${api.config?.ssr ? 'lib' : 'es'}/locale/${getAntdLocale(
+          args.lang,
+          args.country,
+        )}`,
+      ],
+      args,
+    });
+
+  const getList = async (): Promise<IGetLocaleFileListResult[]> => {
     return getLocaleList({
       localeFolder: api.config?.singular ? 'locale' : 'locales',
       separator: api.config.locale?.baseSeparator,
       absSrcPath: paths.absSrcPath,
       absPagesPath: paths.absPagesPath,
+      addAntdLocales,
     });
   };
 
@@ -63,31 +83,54 @@ export default (api: IApi) => {
   api.addRuntimePluginKey(() => 'locale');
 
   // 生成临时文件
-  api.onGenerateFiles(() => {
+  api.onGenerateFiles(async () => {
     const localeTpl = readFileSync(
-      join(__dirname, 'templates', 'locale.tpl'),
+      join(winPath(__dirname), 'templates', 'locale.tpl'),
       'utf-8',
     );
-    const { baseSeparator, baseNavigator, antd, title } = api.config
-      .locale as ILocaleConfig;
+    const { baseSeparator, baseNavigator, antd, title, useLocalStorage } = api
+      .config.locale as ILocaleConfig;
     const defaultLocale = api.config.locale?.default || `zh${baseSeparator}CN`;
 
-    const localeList = getList();
+    const localeList = await getList();
     const momentLocales = localeList
       .map(({ momentLocale }) => momentLocale)
       .filter(locale => locale);
-
     const antdLocales = localeList
       .map(({ antdLocale }) => antdLocale)
       .filter(locale => locale);
 
+    let MomentLocales = momentLocales;
+    let DefaultMomentLocale = '';
+    // set moment default accounding to locale.default
+    if (!MomentLocales.length && api.config.locale?.default) {
+      const [lang, country = ''] = defaultLocale.split(baseSeparator);
+      const { momentLocale } = getMomentLocale(lang, country);
+      if (momentLocale) {
+        MomentLocales = [momentLocale];
+        DefaultMomentLocale = momentLocale;
+      }
+    }
+
+    let DefaultAntdLocales: string[] = [];
+    // set antd default locale
+    if (!antdLocales.length && api.config.locale?.antd) {
+      const [lang, country = ''] = defaultLocale.split(baseSeparator);
+      DefaultAntdLocales = lodash.uniq(
+        await addAntdLocales({
+          lang,
+          country,
+        }),
+      );
+    }
+
     api.writeTmpFile({
       content: Mustache.render(localeTpl, {
-        MomentLocales: momentLocales,
-        UseSSR: !!api.config?.ssr,
+        MomentLocales,
+        DefaultMomentLocale,
+        DefaultAntdLocales,
         Antd: !!antd,
         Title: title && api.config.title,
-        AntdLocales: antdLocales,
         BaseSeparator: baseSeparator,
         DefaultLocale: defaultLocale,
         DefaultLang: defaultLocale,
@@ -104,8 +147,9 @@ export default (api: IApi) => {
       content: Mustache.render(localeExportsTpl, {
         BaseSeparator: baseSeparator,
         BaseNavigator: baseNavigator,
+        UseLocalStorage: !!useLocalStorage,
+        LocaleDir: api.config.singular ? 'locale' : 'locales',
         LocaleList: localeList,
-        UseSSR: !!api.config?.ssr,
         Antd: !!antd,
         DefaultLocale: JSON.stringify(defaultLocale),
         warningPkgPath: winPath(require.resolve('warning')),
@@ -141,8 +185,8 @@ export default (api: IApi) => {
   );
 
   // watch locale files
-  api.addTmpGenerateWatcherPaths(() => {
-    const localeList = getList();
+  api.addTmpGenerateWatcherPaths(async () => {
+    const localeList = await getList();
     return exactLocalePaths(localeList);
   });