+++ packages/plugin-access/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.3.0](https://github.com/umijs/plugins/compare/@umijs/plugin-access@2.2.1...@umijs/plugin-access@2.3.0) (2020-03-24)
+
+### Features
+
+- Make parent route unaccessible if child routes exist and all of child routes are unaccessible ([#105](https://github.com/umijs/plugins/issues/105)) ([f4e4d48](https://github.com/umijs/plugins/commit/f4e4d48f62418b6caa7998dc24954c835776fcb0))
+
+## [2.2.1](https://github.com/umijs/plugins/compare/@umijs/plugin-access@2.2.0...@umijs/plugin-access@2.2.1) (2020-03-18)
+
+### Bug Fixes
+
+- access hmr ([baacdf2](https://github.com/umijs/plugins/commit/baacdf22bf84682c90698d722866aa8fe6f8edb9))
+
 # [2.2.0](https://github.com/umijs/plugins/compare/@umijs/plugin-access@2.1.1...@umijs/plugin-access@2.2.0) (2020-02-18)
 
 ### Features