+++ packages/preset-ant-design-pro/CHANGELOG.md
@@ -2,6 +2,22 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.2](https://github.com/umijs/plugins/compare/@umijs/preset-ant-design-pro@1.1.1...@umijs/preset-ant-design-pro@1.1.2) (2020-03-18)
+
+**Note:** Version bump only for package @umijs/preset-ant-design-pro
+
+## [1.1.1](https://github.com/umijs/plugins/compare/@umijs/preset-ant-design-pro@1.1.0...@umijs/preset-ant-design-pro@1.1.1) (2020-03-18)
+
+### Bug Fixes
+
+- **preset-ant-design-pro:** fix plugin no work bugs ([#111](https://github.com/umijs/plugins/issues/111)) ([dd90a54](https://github.com/umijs/plugins/commit/dd90a54495d448e6d674945cf0305e99685b352d))
+
+# [1.1.0](https://github.com/umijs/plugins/compare/@umijs/preset-ant-design-pro@1.0.1...@umijs/preset-ant-design-pro@1.1.0) (2020-03-18)
+
+### Features
+
+- add umi-plugin-pro-block pro ([#104](https://github.com/umijs/plugins/issues/104)) ([9e52138](https://github.com/umijs/plugins/commit/9e52138bc5be752f75edb059b349e5afd92f890d))
+
 ## [1.0.1](https://github.com/umijs/plugins/compare/@umijs/preset-ant-design-pro@1.0.0...@umijs/preset-ant-design-pro@1.0.1) (2020-03-13)
 
 **Note:** Version bump only for package @umijs/preset-ant-design-pro